[Editorial] Replace italics with quotes.
Originally I read the italics as marking "élan" as a French borrowing,
despite this being a familiar English word by the 20th century. But a
closer reading reveals that he is using "élan" as a technical term
specific to the visual arts. SEMOS 8.2.10 indicates that this should
therefore be quoted instead of italicized.