Remove italics on Follies when not an actual performance.
The idea is that "Follies" refers to both the theatrical troupe and to
the musical revue that the troupe performs. The latter should be
italicized, the former should not. It's a bit confusing because some
sentences could be read equally well either way. (I deliberately chose
to have one of each form appear in Chapter 1, so that the reader is
introduced to the distinction right off, and doesn't see it as a
typo when it crops up again in Chapter 15.)